Phase 2 | Piyushkumar | generated code for SQL queries using sqlc and intergrated that in repository

+++ internal/domain/transfer.go
@@ -10,14 +10,12 @@ const (
 	TransferFailed    TransferStatus = "FAILED"
 )
 
-// Transfer.Amount is stored in the smallest currency unit (cents), matching
-// the BIGINT column in the transfers table.
 type Transfer struct {
 	ID             string
 	IdempotencyKey string
 	FromWalletID   string
 	ToWalletID     string
-	Amount         int64
+	Amount         float64
 	Status         TransferStatus
 	FailureReason  *string
 	CreatedAt      time.Time
